requirements.txt
Django==3.1.7
mysqlclient==2.0.3
djangorestframework==3.12.4
torchvision==0.9.1
MindsDB==2.24.1
